Simplify .gitignore to include a single CMake binary directory
Because in-source builds are so widely accepted as a bad idea, it's
reasonable to ignore a single binary directory and nothing else. If
users choose to dump their build products into the repo then they
may do so at their own risk. For users who put their binary directory
outside of their source directory then this is not even necessary
so we only need to cater to users who want the convenience of cloning
a repo and immediately running `cmake -B build`.