Don't over-specify language requirements
Disabling compiler extensions isn't a bad idea because it changes
-std=gnu++17 to -std=c++17 and it makes it harder to accidentally
use extensions but it's not really particularly helpful given how
hard it is to accidentally use compiler extensions. Besides, the CI
pipeline already in this project will build with enough compilers
to guarantee that compilers extensions are not being used.

As for CMAKE_CXX_STANDARD_REQUIRED, this one is overspecifying the
requirements. If you request C++17 but only use C++14 features, it's
fine if your compiler only supports C++14. It's better to let a
compiler attempt to build the project than fail prematurely before
even giving it a chance to work.