Created BackgroundGrid.cpp
This file should contain all the logic needed to run a quick 2D Grid for the game as a 1D vector internally. It contains methods to check squares, change squares values, see if all aliens (Value 1) have been defeated, is completely scalable thanks to its constructor. To use it, create an object of it with Constructor BackgroundGrid(int x, int y), where x and y are how big you want the grid to be. The other methods all scale with this, and throw errors if an out of bounds access request occurs.